Adiciona dependências do Koin e Retrofit Adiciona dependencia do Moshi Configura buildConfig para esconder a URL e o token Implementa ViewBinding Configura o Koin para o modulo de Network Configura o Retrofit
