+++ app/build.gradle.kts
@@ -17,7 +17,6 @@ android {
     defaultConfig {
         applicationId = "com.hacka.presenteperfeito"
         minSdk = 26
-        //noinspection EditedTargetSdkVersion
         targetSdk = 35
         versionCode = 1
         versionName = "1.0"
@@ -30,11 +29,9 @@ android {
 
     buildTypes {
         debug {
-            buildConfigField("String", "API_TOKEN", "\"${localProperties["API_TOKEN"]}\"")
             buildConfigField("String", "BASE_URL", "\"${localProperties["BASE_URL"]}\"")
         }
         release {
-            buildConfigField("String", "API_TOKEN", "\"${localProperties["API_TOKEN"]}\"")
             buildConfigField("String", "BASE_URL", "\"${localProperties["BASE_URL"]}\"")
 
             isMinifyEnabled = false
@@ -64,7 +61,6 @@ android {
     }
 
     buildFeatures {
-        viewBinding = true
         buildConfig = true
     }
 }
@@ -78,8 +74,12 @@ dependencies {
     implementation(libs.koin.compose.viewmodel)
 
     implementation(libs.retrofit)
-    implementation(libs.converter.gson)
     implementation(libs.okhttp)
+    implementation(libs.logging.interceptor)
+
+    implementation(libs.moshi)
+    implementation(libs.converter.moshi)
+    implementation(libs.moshi.kotlin)
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
