+++ app/src/main/java/com/hacka/presenteperfeito/PerfectGiftApp.kt
@@ -1,10 +1,10 @@
 package com.hacka.presenteperfeito
 
 import android.app.Application
-import com.hacka.presenteperfeito.core.common.networkModule
+import com.hacka.presenteperfeito.core.common.newtwork.networkModule
 import org.koin.core.context.GlobalContext.startKoin
 
-class MyApp : Application() {
+class PerfectGiftApp : Application() {
     override fun onCreate() {
         super.onCreate()
