--- app/src/main/java/com/hacka/presenteperfeito/core/common/ChatApiService.kt
@@ -1,4 +1,0 @@
-package com.hacka.presenteperfeito.core.common
-
-interface ChatApiService {
-}
