+++ app/src/main/java/com/hacka/presenteperfeito/core/common/newtwork/ChatApiService.kt
@@ -1,0 +1,4 @@
+package com.hacka.presenteperfeito.core.common.newtwork
+
+interface ChatApiService {
+}
