+++ app/src/main/java/com/hacka/presenteperfeito/core/common/newtwork/NetworkModule.kt
@@ -1,0 +1,43 @@
+package com.hacka.presenteperfeito.core.common.newtwork
+
+import com.hacka.presenteperfeito.BuildConfig
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import org.koin.dsl.module
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+import java.util.concurrent.TimeUnit
+
+const val TIMEOUT = 60L
+
+val networkModule = module {
+
+    single<OkHttpClient> {
+        val builder = OkHttpClient.Builder()
+
+
+        val loggingInterceptor = HttpLoggingInterceptor().apply {
+            level = HttpLoggingInterceptor.Level.BODY
+        }
+        builder.addInterceptor(loggingInterceptor)
+
+
+        builder
+            .connectTimeout(TIMEOUT, TimeUnit.SECONDS)
+            .readTimeout(TIMEOUT, TimeUnit.SECONDS)
+            .writeTimeout(TIMEOUT, TimeUnit.SECONDS)
+            .build()
+    }
+
+    single<Retrofit> {
+        Retrofit.Builder()
+            .baseUrl(BuildConfig.BASE_URL)
+            .client(get())
+            .addConverterFactory(MoshiConverterFactory.create())
+            .build()
+    }
+
+    single<ChatApiService> {
+        get<Retrofit>().create(ChatApiService::class.java)
+    }
+}
